feat: 해설 순서 인덱스 필드 추가

+++ backend/api-server/storage/db-mongo/src/main/java/com/technote/storage/mongo/core/Note.java
@@ -39,14 +39,14 @@ public class Note {
     @Getter
     @Builder
     public static class Commentary {
-        private double startTime;
+        private int startTime;
         private String content;
     }
 
     @Getter
     @Builder
     public static class Segment {
-        private double startTime;
+        private int startTime;
         private String title;
         private String summary;
     }
